Make RESTClientTest Inherit from RESTClient
Conceptually RESTClientTest is a RESTClient, just one intended for testing. This means that isinstance(obj, RESTClient) type checks will pass with RESTClientTest. More importantly for me anyway ;-) while the super call in the __init__ is sufficient for most client testing, if like in workqueue.Worker you have multiple inheritance you may want to call the __init__ directly rather than use some horrible super fudge with variable args passing around. In this case it currently fails to initialise as the base (RESTClientTest) is not a RESTClient.

	modified:   src/pdm/framework/RESTClient.py