Refactor service handling of worker output.
This fixes https://github.com/ic-hep/pdm/issues/256

This also refactors the copying to fix a hidden bug which would have meant job element ids were not sequential but would have had gaps. The gaos would have arisen when the is dir condition was satisfied and the loop continued to the next element. The patch also refactors the logic of giving an absolute filename to copy to.

	modified:   src/pdm/workqueue/WorkqueueService.py