Fix expansion of COPY jobs for gfal relative dirs stat bug.
When stating relative dirs '.' and '..' gfal reports these with st_mode that evaluates as a regular file rather than a directory ie. S_ISREG = True, S_ISDIR = False. This is a quick dirty hack to explicitly exclude these dirs from COPY job expansion. It should not be necessary and when gfal is fixed we can revert this.

Note also a small change to logging to make debugging easier.

	modified:   src/pdm/workqueue/WorkqueueService.py