Rename the force_fail method to force_complete.
This change better reflects the fact that it is the workqueue not the worker that decides if an element is failed based on the return code. This method merely forces the Worker to return all outstanding elements to the workqueue with a given returncode (Which could in theory be a successful 0). Also fixed the missing docstring.

	modified:   src/pdm/workqueue/Worker.py